+++ orca/cli/commands.py
@@ -21,5 +21,6 @@ def run(file, verbose, args):
     conf = process_config(file)
     print('process run ' + json.dumps(conf, indent=2))
     config = OrcaConfig(conf, file.name, args)
-    config.execute()
-
+    
+    handler = ExecutionHandler()
+    handler.execute(config)
