+++ orca/core/tasks.py
@@ -1,10 +1,6 @@
 import logging
 from typing import Dict, List
-from csip import Client
 from dotted.collection import DottedCollection
-import requests
-import subprocess
-import os
 
 log = logging.getLogger(__name__)
 
@@ -55,117 +51,3 @@ class OrcaTask(object):
     @property
     def python(self) -> str:
         return self.task_data.get('python')
-
-
-def values_tostr(d: Dict) -> Dict:
-    a = {key: str(value) for key, value in d.items()}
-    return a
-
-
-def handle_service_result(response: Dict, outputs, name: str) -> Dict:
-    d = dict()
-    for k, v in response.items():
-        print(" handling req result props: {0} -> {1}".format(k, str(v)))
-        if k in outputs:
-            d[name + "." + k] = v
-    return d
-
-
-def handle_csip_result(response: Dict, outputs: List, name: str) -> Dict:
-    d = dict()
-    for k,v in response.items():
-        if k in outputs:
-            d[name + "." + k] = v['value']
-    return d
-
-
-def handle_python_result(outputs: List, name: str)-> Dict:
-    d = dict()
-    for v in outputs:
-        d[name + '.' + v] = eval(v, globals())
-
-    return d
-
-
-def handle_csip(task: OrcaTask, yaml_dir: str) -> Dict:
-    try:
-        url = task.csip
-        name = task.name
-        outputs = task.outputs
-        client = Client()
-        for key, value in task.inputs.items():
-            if isinstance(value, DottedCollection):
-                client.add_data(key, value.to_python())
-
-            else:
-                client.add_data(key, value)
-        client = client.execute(url)
-        return handle_csip_result(client.data, outputs, name)
-    except requests.exceptions.HTTPError as e:
-        raise OrcaTaskException(e)
-
-
-def handle_http(task: OrcaTask, yaml_dir: str) -> Dict:
-    url = task.http
-    name = task.name
-    inputs = task.inputs
-    if 'method' not in task.config:
-        raise OrcaTaskException("requests service operator must include method: service {0}".format(name))
-    if task.config['method'] == 'GET':
-        return handle_service_result(requests.get(url, params=task.config['params']).content, name)
-    elif task.config['method'] == 'POST':
-        if isinstance(inputs, DottedCollection):
-            return handle_service_result(requests.post(url, inputs.to_python()).content, name)
-        else:
-            return handle_service_result(requests.post(url, inputs).content, name)
-
-
-def handle_bash(task: OrcaTask, yaml_dir: str):
-    env = {}
-    cmd = task.bash
-    config = task.config
-    # get defaults
-    delimiter = config.get('delimiter', '\n')
-    errmsg = config.get('errmsg', "unspecified error")
-    exit_on_err = config.get("exitOnError", True)
-    wd = config.get('wd', None)
-
-    if len(task.inputs) > 0:
-        env = values_tostr(task.inputs)
-    sp = subprocess.Popen(cmd, env=env, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=wd)
-    out, err = sp.communicate()
-    if err:
-        for line in err.split('\n'):
-            print("ERROR: " + line)
-    if sp.returncode != 0 and exit_on_err is True:
-        print('ERROR: ' + str(sp.returncode) + " " + errmsg)
-    if out:
-        o = ""
-        for line in out.decode('utf-8').split(delimiter):
-            if line:
-                o += line + '\n'
-        return o
-    return {}
-
-
-def handle_python(task: OrcaTask, yaml_dir: str):
-    print("  exec python file : " + task.python)
-    for k, v in task.inputs.items():
-        if isinstance(v, str):
-            v = "\'{0}\'".format(v)
-        fmt = '{0} = {1}'.format(k, v)
-        exec(fmt, locals(), globals())
-
-    if os.path.isabs(task.python) and os.path.isfile(task.python):
-        # absolute path provided
-        exec(open(task.python).read(), globals())
-    else:
-        if not os.path.isabs(task.python) and os.path.isfile(os.path.join(yaml_dir, task.python)):
-            # path relative to yaml file
-            exec(open(os.path.join(yaml_dir, task.python)).read(), globals())
-        else: 
-            # python inlined
-            exec(task.python, globals())
-        
-    return handle_python_result(task.outputs, task.name)
-
